+++ Modules/Inventory/database/migrations/2025_02_10_002031_create_warehouses_table.php
@@ -11,6 +11,10 @@ return new class extends Migration
      */
     public function up(): void
     {
+        if (Schema::hasTable('warehouses')) {
+            return;
+        }
+
         Schema::create('warehouses', function (Blueprint $table) {
             $table->id();
             $table->foreignId('site_id')->constrained('sites')->onDelete('cascade');
