+++ Modules/Inventory/database/migrations/2025_02_10_002537_create_stock_table.php
@@ -11,6 +11,10 @@ return new class extends Migration
      */
     public function up(): void
     {
+        if (Schema::hasTable('stock')) {
+            return;
+        }
+
         Schema::create('stock', function (Blueprint $table) {
             $table->id();
             $table->foreignId('product_id')->constrained('products')->onDelete('cascade');
