+++ Modules/Inventory/database/migrations/2025_02_10_002634_create_product_instances_table.php
@@ -11,6 +11,10 @@ return new class extends Migration
      */
     public function up(): void
     {
+        if (Schema::hasTable('product_instances')) {
+            return;
+        }
+
         Schema::create('product_instances', function (Blueprint $table) {
             $table->id();
             $table->foreignId('product_id')->constrained('products')->onDelete('cascade');
