+++ Modules/Inventory/database/migrations/2025_02_10_003433_create_raw_material_uom_conversions_table.php
@@ -11,6 +11,10 @@ return new class extends Migration
      */
     public function up(): void
     {
+        if (Schema::hasTable('raw_material_uom_conversions')) {
+            return;
+        }
+
         Schema::create('raw_material_uom_conversions', function (Blueprint $table) {
             $table->id();
             $table->foreignId('from_uom_id')->constrained('item_uoms')->onDelete('cascade');
