+++ Modules/Production/database/migrations/2025_07_15_062741_create_routing_steps_table.php
@@ -11,6 +11,10 @@ return new class extends Migration
      */
     public function up(): void
     {
+        if (Schema::hasTable('routing_steps')) {
+            return;
+        }
+
         Schema::create('routing_steps', function (Blueprint $table) {
             $table->id();
             $table->foreignId('routing_id')->constrained()->onDelete('cascade');
